refactor: bug-bounty page to ssr
